Fix migrations

+++ database/migrations/2023_02_04_094500_add_columns_for_dna_matchings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('dna_matchings', function (Blueprint $table) {
-            $table->unsignedInteger('match_id')->nullable();
+            $table->unsignedBigInteger('match_id')->nullable();
         });
     }
